ClientOptions -> Options and moved code from workerclient/client.go to workerclient/workerclient.go
