+++ pkg/workerapi/workerclient/client_test.go
@@ -7,7 +7,7 @@ import (
 )
 
 func TestNew(t *testing.T) {
-	c, err := New("http://127.0.0.1", ClientOptions{InsecureSkipVerify: true})
+	c, err := New("http://127.0.0.1", Options{InsecureSkipVerify: true})
 	assert.NoError(t, err)
 	assert.NotNil(t, c)
 }
