+++ pkg/workerapi/workerclient/grpcclient_test.go
@@ -7,13 +7,13 @@ import (
 )
 
 func TestOpenBadAddress(t *testing.T) {
-	c := newGRPCClient("badaddress", ClientOptions{InsecureSkipVerify: false})
+	c := newGRPCClient("badaddress", Options{InsecureSkipVerify: false})
 	assert.Error(t, c.ensureOpen())
 	assert.NoError(t, c.close())
 }
 
 func TestOpeningTwiceGivesSameConnection(t *testing.T) {
-	c := newGRPCClient("127.0.0.1", ClientOptions{InsecureSkipVerify: true})
+	c := newGRPCClient("127.0.0.1", Options{InsecureSkipVerify: true})
 	assert.NoError(t, c.ensureOpen())
 	conn := c.conn
 	assert.NoError(t, c.ensureOpen())
@@ -22,7 +22,7 @@ func TestOpeningTwiceGivesSameConnection(t *testing.T) {
 }
 
 func TestClosing(t *testing.T) {
-	c := newGRPCClient("127.0.0.1", ClientOptions{InsecureSkipVerify: true})
+	c := newGRPCClient("127.0.0.1", Options{InsecureSkipVerify: true})
 	assert.NoError(t, c.ensureOpen())
 	assert.NotNil(t, c.conn)
 	assert.NoError(t, c.close())
