+++ pkg/workerapi/workerclient/workerclient.go
@@ -4,6 +4,7 @@ package workerclient
 
 import (
 	"context"
+	"fmt"
 	"io"
 
 	v1 "github.com/iver-wharf/wharf-cmd/api/workerapi/v1"
@@ -12,25 +13,6 @@ import (
 )
 
 var log = logger.NewScoped("WORKER-CLIENT")
-
-// Client is an interface with methods to communicate with a Wharf worker server.
-type Client interface {
-	StreamLogs(ctx context.Context, req *LogsRequest, opts ...grpc.CallOption) (v1.Worker_StreamLogsClient, error)
-	StreamStatusEvents(ctx context.Context, req *StatusEventsRequest, opts ...grpc.CallOption) (v1.Worker_StreamStatusEventsClient, error)
-	StreamArtifactEvents(ctx context.Context, req *ArtifactEventsRequest, opts ...grpc.CallOption) (v1.Worker_StreamArtifactEventsClient, error)
-	DownloadArtifact(artifactID uint) (io.ReadCloser, error)
-
-	Close() error
-}
-
-// ClientOptions contains options that can be used in the creation
-// of a new client.
-type ClientOptions struct {
-	// InsecureSkipVerify disables cert verification if set to true.
-	//
-	// Should NOT be true in a production environment.
-	InsecureSkipVerify bool
-}
 
 // LogLine is an alias for workerapi/v1.StreamLogsResponse.
 type LogLine = v1.StreamLogsResponse
@@ -49,3 +31,81 @@ type ArtifactEvent = v1.StreamArtifactEventsResponse
 
 // ArtifactEventsRequest is an alias for workerapi/v1.StreamArtifactEventsResponse.
 type ArtifactEventsRequest = v1.StreamArtifactEventsRequest
+
+// Client is an interface with methods to communicate with a Wharf worker server.
+type Client interface {
+	StreamLogs(ctx context.Context, req *LogsRequest, opts ...grpc.CallOption) (v1.Worker_StreamLogsClient, error)
+	StreamStatusEvents(ctx context.Context, req *StatusEventsRequest, opts ...grpc.CallOption) (v1.Worker_StreamStatusEventsClient, error)
+	StreamArtifactEvents(ctx context.Context, req *ArtifactEventsRequest, opts ...grpc.CallOption) (v1.Worker_StreamArtifactEventsClient, error)
+	DownloadArtifact(artifactID uint) (io.ReadCloser, error)
+
+	Close() error
+}
+
+// New creates a new client that can communicate with a Wharf worker server.
+//
+// Implements the Closer interface.
+func New(baseURL string, opts Options) (Client, error) {
+	rest, err := newRestClient(opts)
+	if err != nil {
+		return nil, err
+	}
+	return &client{
+		rest:    rest,
+		grpc:    newGRPCClient(baseURL, opts),
+		baseURL: baseURL,
+	}, nil
+}
+
+type client struct {
+	rest    *restClient
+	grpc    *grpcClient
+	baseURL string
+}
+
+// Options contains options that can be used in the creation
+// of a new client.
+type Options struct {
+	// InsecureSkipVerify disables cert verification if set to true.
+	//
+	// Should NOT be true in a production environment.
+	InsecureSkipVerify bool
+}
+
+// StreamLogs returns a stream that will receive log lines from the worker.
+func (c *client) StreamLogs(ctx context.Context, req *LogsRequest, opts ...grpc.CallOption) (v1.Worker_StreamLogsClient, error) {
+	if err := c.grpc.ensureOpen(); err != nil {
+		return nil, err
+	}
+	return c.grpc.client.StreamLogs(ctx, req, opts...)
+}
+
+// StreamStatusEvents returns a stream that will receive status events from the
+// worker.
+func (c *client) StreamStatusEvents(ctx context.Context, req *StatusEventsRequest, opts ...grpc.CallOption) (v1.Worker_StreamStatusEventsClient, error) {
+	if err := c.grpc.ensureOpen(); err != nil {
+		return nil, err
+	}
+	return c.grpc.client.StreamStatusEvents(ctx, req, opts...)
+}
+
+// StreamArtifactEvents returns a stream that will receive status events from the
+// worker.
+func (c *client) StreamArtifactEvents(ctx context.Context, req *ArtifactEventsRequest, opts ...grpc.CallOption) (v1.Worker_StreamArtifactEventsClient, error) {
+	if err := c.grpc.ensureOpen(); err != nil {
+		return nil, err
+	}
+	return c.grpc.client.StreamArtifactEvents(ctx, req, opts...)
+}
+
+func (c *client) DownloadArtifact(artifactID uint) (io.ReadCloser, error) {
+	res, err := c.rest.get(fmt.Sprintf("%s/api/artifact/%d/download", c.baseURL, artifactID))
+	if err := assertResponseOK(res, err); err != nil {
+		return nil, err
+	}
+	return res.Body, nil
+}
+
+func (c *client) Close() error {
+	return c.grpc.close()
+}
